services/agent_service/requirements.txt
chromadb
kafka-python
PyPDF2
langchain-groq
sentence-transformers
